# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a table (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Table created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get table by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update a table (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableUpdate"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /tables/by-token/{token}:
    get:
      operationId: getTableByToken
      tags: [tables]
      summary: Get table by QR token (public, client use)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── ORDERS ────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, refused, ready, delivered]
        - name: tableId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithItems"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (client)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (vendor/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
  /orders/table/{tableId}/active:
    get:
      operationId: getActiveOrderByTable
      tags: [orders]
      summary: Get active order for a table (client polling)
      parameters:
        - name: tableId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Active order or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithItems"
        "404":
          description: No active order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── RATINGS ───────────────────────────────────────────────────────────────
  /ratings:
    get:
      operationId: listRatings
      tags: [ratings]
      summary: List ratings (admin/vendor)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"
    post:
      operationId: createRating
      tags: [ratings]
      summary: Submit a rating for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "201":
          description: Rating created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── STATS / DASHBOARD ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Dish:
      type: object
      required: [id, categoryId, nameAr, nameFr, nameEn, price, isAvailable, isPopular, isNew]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        nameAr:
          type: string
        nameFr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: ["string", "null"]
        descriptionFr:
          type: ["string", "null"]
        descriptionEn:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        allergens:
          type: array
          items:
            type: string
        isPopular:
          type: boolean
        isNew:
          type: boolean
        isAvailable:
          type: boolean
...
    Table:
      type: object
      required: [id, number, qrToken, isActive]
      properties:
        id:
          type: integer
        number:
          type: integer
        qrToken:
          type: string
        isActive:
          type: boolean
        status:
          type: string
          enum: [free, occupied, waiting]
    TableInput:
      type: object
      required: [number]
      properties:
        number:
          type: integer
    TableUpdate:
      type: object
      properties:
        number:
          type: integer
        isActive:
          type: boolean

    # ── ORDERS ──────────────────────────────────────────────────────────────
    OrderItem:
      type: object
      required: [id, dishId, quantity, unitPrice]
      properties:
        id:
          type: integer
        dishId:
          type: integer
        dish:
          $ref: "#/components/schemas/Dish"
        quantity:
          type: integer
        unitPrice:
          type: number
        customNote:
          type: ["string", "null"]
    OrderWithItems:
      type: object
      required: [id, tableId, status, totalPrice, createdAt, updatedAt, items]
      properties:
        id:
          type: integer
        tableId:
          type: integer
        table:
          $ref: "#/components/schemas/Table"
        status:
          type: string
          enum: [pending, confirmed, refused, ready, delivered]
        totalPrice:
          type: number
        note:
          type: ["string", "null"]
        refusalReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        rating:
          $ref: "#/components/schemas/Rating"
    OrderInput:
      type: object
      required: [tableId, items]
      properties:
        tableId:
          type: integer
        note:
          type: string
        items:
          type: array
          items:
            type: object
            required: [dishId, quantity]
            properties:
              dishId:
                type: integer
              quantity:
                type: integer
              customNote:
                type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, refused, ready, delivered]
        refusalReason:
          type: string

    # ── RATINGS ─────────────────────────────────────────────────────────────
    Rating:
      type: object
      required: [id, orderId, stars, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RatingInput:
      type: object
      required: [orderId, stars]
      properties:
        orderId:
          type: integer
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string

    # ── STATS ────────────────────────────────────────────────────────────────
...